Ht test

+++ src/main/java/hashTable/HashTable.java
@@ -15,7 +15,7 @@ public class HashTable<T> {
     public int getHash(String key){
 
         //get char code sum
-        int index = 0;
+        long index = 0;
         for (char c : key.toCharArray()){
             index += (int) c;
         }
@@ -24,7 +24,7 @@ public class HashTable<T> {
         index *= 937;
         index = index % this.buckets.length;
 
-        return index;
+        return (int) index;
     }
 
     //This method adds a node to the hashtable
@@ -35,6 +35,7 @@ public class HashTable<T> {
 
         //add the node to table with index from hash
         Node<T> nodeToAdd = new Node<>(key, value);
+
         if (this.buckets[index] == null){
             this.buckets[index] = new LinkedList<>();
         }
